주문하고 주문목록 조회까지 학습

+++ order-service/src/main/java/com/green/eats/order/application/model/OrderPostReq.java
@@ -1,12 +1,8 @@
 package com.green.eats.order.application.model;
 
-import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import java.util.List;
 
@@ -16,5 +12,5 @@ public class OrderPostReq {
   private List<OrderPostItemReq> items;
 
   @Positive(message = "0 이상이어야 합니다.")
-  private Long totalAmount;
+  private Integer totalAmount;
 }
